requirements.txt
argparse==1.2.1
GitPython==2.1.0
glob2==0.5
httplib2==0.8
imageio==2.1.2
matplotlib==1.5.3
MIDIUtil==1.1.3
moviepy==0.2.3.1
msgpack-python==0.4.8
numpy==1.11.2
pandas==0.13.1
pip==1.5.4
pretty-midi==0.2.7
pyFluidSynth==1.2.5
python-midi==v0.2.4
requests==2.12.1
scikit-learn==0.18.1
scipy==0.13.3
six==1.10.0
sklearn==0.0
tensorflow==0.10.0
tqdm==4.10.0
virtualenv==1.11.4
